rename files and paths for shared/detail
