+++ feature-libs/organization/administration/components/cost-center/details/cost-center-details.component.spec.ts
@@ -5,7 +5,7 @@ import { CostCenter, I18nTestingModule } from '@spartacus/core';
 import { UrlTestingModule } from 'projects/core/src/routing/configurable-routes/url-translation/testing/url-testing.module';
 import { of, Subject } from 'rxjs';
 import { CardTestingModule } from '../../shared/card/card.testing.module';
-import { ToggleStatusModule } from '../../shared/organization-detail/toggle-status-action/toggle-status.module';
+import { ToggleStatusModule } from '../../shared/detail/toggle-status-action/toggle-status.module';
 import { OrganizationItemService } from '../../shared/organization-item.service';
 import { MessageTestingModule } from '../../shared/organization-message/message.testing.module';
 import { MessageService } from '../../shared/organization-message/services/message.service';
