+++ feature-libs/organization/administration/components/shared/detail/index.ts
