+++ feature-libs/organization/administration/components/user/details/user-details.module.ts
@@ -4,7 +4,7 @@ import { RouterModule } from '@angular/router';
 import { I18nModule, UrlModule } from '@spartacus/core';
 import { ItemExistsModule } from '../../shared/item-exists.module';
 import { CardModule } from '../../shared/card/card.module';
-import { ToggleStatusModule } from '../../shared/organization-detail/toggle-status-action/toggle-status.module';
+import { ToggleStatusModule } from '../../shared/detail/toggle-status-action/toggle-status.module';
 import { UserDetailsComponent } from './user-details.component';
 
 @NgModule({
